+++ lizmap/saas.py
@@ -17,7 +17,7 @@ from lizmap.project_checker_tools import _is_vector_pg
 from lizmap.qgis_plugin_tools.tools.i18n import tr
 
 
-def is_lizmap_dot_com_hosting(metadata: dict) -> bool:
+def is_lizmap_cloud(metadata: dict) -> bool:
     """ Return True if the metadata is coming from lizmap.com. """
     if not metadata:
         # Mainly in tests?